*** empty log message ***
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@1813 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ M2/Macaulay2/e/mem.cpp
@@ -234,7 +234,7 @@ extern "C" void outofmem();
 
 // the new way:
 
-void* operator new( unsigned int size ) {
+void* operator new( size_t size ) {
   void *p = GC_MALLOC( size );
   if (p == NULL) outofmem();
   return p;
@@ -244,7 +244,7 @@ void operator delete( void* obj ) {
   if (obj != NULL) GC_FREE( obj );
 }
 
-void* operator new []( unsigned int size ) {
+void* operator new []( size_t size ) {
   void *p = GC_MALLOC( size );
   if (p == NULL) outofmem();
   return p;
